backend/requirements.txt
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.6.2.post1
APScheduler==3.11.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
bcrypt==4.2.1
bidict==0.23.1
blinker==1.9.0
CacheControl==0.14.1
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
cryptography==44.0.0
distro==1.9.0
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
fastapi==0.115.6
filelock==3.17.0
firebase-admin==6.6.0
Flask==3.1.0
fsspec==2024.12.0
gevent==24.11.1
gevent-websocket==0.10.1
google-api-core==2.24.0
google-api-python-client==2.157.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.19.0
google-cloud-storage==2.19.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
greenlet==3.1.1
grpcio==1.69.0
grpcio-status==1.69.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.27.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.7.1
macholib==1.16.3
MarkupSafe==3.0.2
motor==3.6.1
mpmath==1.3.0
msgpack==1.1.0
networkx==3.4.2
nltk==3.9.1
openai==1.60.1
packaging==24.2
passlib==1.7.4
pefile==2023.2.7
pillow==11.1.0
pipdeptree==2.24.0
proto-plus==1.25.0
protobuf==5.29.2
psutil==6.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.4
pydantic_core==2.27.2
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2024.10
PyJWT==2.10.1
pymongo==4.9.2
pyparsing==3.2.1
PyQt6==6.8.0
PyQt6-Qt6==6.8.1
PyQt6_sip==13.9.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.10.1
python-jose==3.3.0
python-json-logger==2.0.7
python-multipart==0.0.19
python-socketio==5.11.4
pywin32-ctypes==0.2.3
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rsa==4.9
simple-websocket==1.1.0
six==1.16.0
sniffio==1.3.1
spacy==3.8.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
starlette==0.41.3
sympy==1.13.1
threadpoolctl==3.5.0
tokenizers==0.21.0
tqdm==4.67.0
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.32.0
websocket-client==1.8.0
Werkzeug==3.1.3
wsproto==1.2.0
zope.event==5.0
zope.interface==7.1.1
yake==0.4.8